Change parser logic to allow for floating point numbers

+++ src/com/runaway/model/Parser.java
@@ -10,6 +10,6 @@ public interface Parser {
 	 * Given a string of numbers, return the sum
 	 * @return the sum of the integers
 	 */
-	public int calculateRow(String string);
+	public double calculateRow(String string);
 
 }
